Bug fixes with regards to the Duration Worked on both Projects and Tasks. Modified some of the view templates to be more responsive / display data more efficiently. Added a jQuery script that uses an Ajax request to process a project or a task's current status.

+++ TimesheetSystem/TimesheetSystem/DAL/Models/TimeLog.cs
@@ -19,7 +19,7 @@ namespace TimesheetSystem.DAL.Models
         public DateTime TimeLogDate { get; set; }
 
         [Display(Name = "Work Duration")]
-        public int DurationWorked { get; set; }
+        public double DurationWorked { get; set; }
 
         public int TasksId { get; set; }
